input: misc: drv260x_input: Rename to qti-haptics
This driver behaves like the qti haptics driver and hence we can
use the qti vibrator HAL. Since [1] the qti vibrator HAL checks
wether a qti haptics device is detected and does not work if it's
not a qti one. In order to allow the HAL to work with this driver
we spoof the input device name to qti-haptics if the actual
qti-haptics driver is not built.

[1] LineageOS/android_vendor_qcom_opensource_vibrator@c8f31f1